+++ API/index.js
@@ -13,7 +13,6 @@ const ObjectId = require('mongodb').ObjectId;
 const crypto = require ('crypto');
 //
 const mongoURL = 'mongodb://localhost:27017/cliente_proveedor/';
-const collName = 'clientes';
 
 const multer = require('multer');
 const {ClientRequest} = require('http');
@@ -24,16 +23,31 @@ const mimeParser = multer({
 		},
 });
 
-var collection;
+var collClien;
+var collProv;
+var collProduc;
 
 async function conectadb () {
 	var client = await MongoClient.connect(mongoURL);
-	collection = await client.db().collection(collName);
+	collClien = await client.db().collection('clientes');
+	collProv = await client.db().collection('proveedores');
+	collProduc = await client.db().collection('productos');
 }
 
 conectadb();
 port = 3000;
 
+/**
+ * #####################################
+ * ############## Inicio de los endpoint
+ * #####################################
+ */
+
+
+ /**
+  * ############# Endpoints clientes ###
+  * ####################################
+  */
 app.post('/nuevoCli/', async (req,res)=>{
 	var nuevoClientes = req.body.clientes; 
 	var pwd = req.body.pwd;
@@ -63,8 +77,8 @@ app.post('/nuevoCli/', async (req,res)=>{
 		file:req.body.file,
 	};
 	//var result = await collFiles.insertOne(file);
-	var mongoRes = await collection.insertOne(documento);
-	var clientes = await collection.find().toArray();
+	var mongoRes = await collClien.insertOne(documento);
+	var clientes = await collClien.find().toArray();
 	var json = JSON.stringify(clientes);
 	//console.log(file);
 	//res.send(result.insertedId);
@@ -76,37 +90,102 @@ app.get('/listadoCli/', async (req,res)=>{
 	/**
 	* Endpoint: http://localhost:3000/listadoCli/
 	*/
-	var clientes = await collection.find().toArray();
+	var clientes = await collClien.find().toArray();
 	var json = JSON.stringify(clientes)
 		res.send(json);		
 })
 
-app.get('/modificarDoc/', async (req,res)=>{
-	/**
-	* Endpoint: http://localhost:3000/borrar/?id=AAAAAAA
-	*/
-	/*var id = req.query.id; 
-	var filtro = {
-		_id: ObjectId(id)
-	}
-    await collection.deleteOne(filtro);*/
-    //quiero q se pueda modificar los atributos del objeto, documento.
-
-	var clientes = await collection.find().toArray();
-	var json = JSON.stringify(clientes)
-	res.send(json);
+/*
+app.get('/modificarFichaCli/', async (req,res)=>{
 })
-app.get('/eliminarUsuario/', async(req, res)=>{
-	var clientes = await collection.find().toArray();
+*/
+
+/*
+app.get('/eliminarCli/', async(req, res)=>{
 	var id = req.query.id;
 	var filtro = {
 		_id:ObjectId(id),
 	}
-	var bajaUsuario = await collection.deleteOne(filtro);
+	var bajaUsuario = await collClien.deleteOne(filtro);
+	var clientes = await collClien.find().toArray();
 	var json = JSON.stringify(clientes);
 	res.send(json);
-	res.send(bajaUsuario);
 })
+*/
+
+
+ /**
+  * ############# Endpoints proveedores 
+  * ####################################
+  */
+
+ /*
+app.post('/nuevoProv/',async (req,res)=>{
+    var nuevoProv= req.body.proveedor;
+    var pwd = req.body.pwd;
+	const hash = crypto.createHash('sha256');
+	hash.update(pwd);
+    var hashString = hash.digest('base64');
+    var datosProv = {
+        proveedor:req.body.proveedor,
+        nifCif:req.body.nifCif,
+        email:req.body.email,
+        clave:hashString,        
+    };
+    var mongoRes = await collection.insertOne(datosProv);
+    var proveedor = await collection.find().toArray();
+    var json = JSON.stringify(proveedor);
+    res.send(json);
+})
+*/
+
+/*
+app.get('/listadoProv/', async (req, res)=>{
+    var proveedor = await collection.find().toArray();
+    var json = JSON.stringify(proveedor);
+        res.send(json);
+})
+*/
+
+/*
+app.post('/listadoProduc/', async (req,res)=>{
+    var productos = {
+        producto:req.body.productos,
+        precio:req.body.precio,
+    }
+    var mongoRes = await collProduc.insertOne(productos);
+    var productos = await collProduc.find().toArray();
+    var json =JSON.stringify(productos);
+        res.send(json);
+})
+*/
+
+/*
+app.get('/modificarProv/', async (req, res)=>{
+    var proveedor = await collection.find().toArray();
+    var json = JSON.stringify(proveedor);
+        res.send(json);
+})
+*/
+
+/*
+app.get('/modificarProduc/',async (req, res)=>{
+    var productos = await collProduc.find().toArray();
+    var json = JSON.stringify(productos);
+        res.send(json);
+})
+*/
+
+
+
+
+
+
+/**
+ * #####################################
+ * ############## Fin de los endpoint
+ * #####################################
+ */
 
 app.listen(port,()=>{
 	console.log(`Todo listo y escuchando en http://localhost:${port}/`);
