+++ Svelte/src/Login.svelte
@@ -1,0 +1,16 @@
+<script>
+    var nick,  clave
+    async function entrarSocio(){
+        for (listadoClientes as cliente)
+            if [(nick===nick)&&(clave===clave)]{
+              //debería llevar a la pag de socios donde elegir la compra
+              (else)
+                <p>Comprueba si el nick y la contraseña son correctas<p>}
+    };
+</script>
+
+<div>
+    <input id="nick" type="text" placeholder="Completa este campos" bind:value={nick}>
+    <input id="clave" type="password" placeholder="Escribe tu clave para poder acceder" bind:value={clave}>
+    <button on:click={entrarSocio()}>Enviar</button>
+</div>
