+++ Svelte/src/PerfilUsuario.svelte
@@ -1,0 +1,12 @@
+<script>
+    async function getModificarDoc(){
+        var getModificarDocURL = 'http://localhost:3000/modificarDoc/';
+        var resp = await fetch(getModificarDocURL);
+        listadoClientes = await resp.json();
+        borrar();
+    }
+</script>
+
+<div>
+    
+</div>
